Added two methods: restore punct and capitals
Added two methods to the script. The method to restore capitals takes an
array of string and an array of indices and capitalizes those indexed
words. Similarly, the method to restore punctuation takes an array of
strings and a hash of punctuation marks (keys) and indices (values as
lists) and punctuates the appropriate words in the returned list. It's
important to note that since these methods rely on indexing, there exists
a certain point the the script after which the number and order of the
words cannot be changed! So I gotta keep that in mind.