Add main function skeleton and update dictionary
Added main function skeleton and updated dictionary to include entries
from universal hub. Discovered that I can do one word to multiple word
conversion, but not multiple word keys, as that would change the indexing
of the array containing the phrase to be translated.

I need to finish the main function by adding a translate option and a word
lookup option. The dictionary also needs to be updated to include plural
nouns, and past/progressive tenses.