Complete main translator method and bug squash
The find and store capitals method couldn't handle multiples of the same capital letter (I), so I had to switch from indexing the string based on the word to a simple counter. I thought that the find and store punctuation method would have to change as well, but since it chomps the punctuation right after it finds it, there are no repeats.

I also added some puts messages and did some line formatting to make the script output more readable. I moved the unused code to the bottom of the script and also moved the notes that I had from starting the project.